+++ test/.eslintrc.js
@@ -1,0 +1,8 @@
+// @ts-check
+
+/** @type {import("eslint").Linter.Config} */
+const testEslintConfig = {
+    extends: ["jacobfischer-jest", "../.eslintrc.js"],
+};
+
+module.exports = testEslintConfig;